+++ galsim/__init__.py
@@ -110,7 +110,9 @@ from .noise import DeviateNoise, VariableGaussianNoise
 from .correlatednoise import CorrelatedNoise, getCOSMOSNoise, UncorrelatedNoise
 
 # GSObject
-from .base import GSParams, GSObject, Gaussian, Moffat, Airy, Kolmogorov, Pixel, Box, TopHat
+from .gsobject import GSObject
+from .gsparams import GSParams
+from .base import Gaussian, Moffat, Airy, Kolmogorov, Pixel, Box, TopHat
 from .base import Exponential, Sersic, DeVaucouleurs, Spergel
 from .real import RealGalaxy, RealGalaxyCatalog, simReal
 from .phase_psf import Aperture, PhaseScreenList, PhaseScreenPSF, OpticalPSF
@@ -126,7 +128,7 @@ from .compound import FourierSqrt, FourierSqrtProfile
 from .transform import Transform, Transformation
 
 # Chromatic
-from .chromatic import ChromaticObject, ChromaticAtmosphere, Chromatic, ChromaticSum
+from .chromatic import ChromaticObject, ChromaticAtmosphere, ChromaticSum
 from .chromatic import ChromaticConvolution, ChromaticDeconvolution, ChromaticAutoConvolution
 from .chromatic import ChromaticAutoCorrelation, ChromaticTransformation
 from .chromatic import ChromaticFourierSqrtProfile